# contracts/source/openapi.yaml
openapi: 3.1.0
info:
  title: SourceHarbor API
  version: 0.1.0
  description: API contract for subscriptions, ingest polling, jobs, videos and artifacts.
servers:
  - url: http://127.0.0.1:9000
paths:
...
  /api/v1/watchlists:
    get:
      operationId: listWatchlists
      responses:
        '200':
          description: Watchlist list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Watchlist'
    post:
      operationId: upsertWatchlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WatchlistUpsertRequest'
      responses:
        '200':
          description: Watchlist upsert result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Watchlist'
  /api/v1/watchlists/{watchlist_id}:
    delete:
      operationId: deleteWatchlist
      parameters:
        - in: path
          name: watchlist_id
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
        '404':
          description: Not found
  /api/v1/watchlists/{watchlist_id}/trend:
    get:
      operationId: getWatchlistTrend
      parameters:
        - in: path
          name: watchlist_id
          required: true
          schema:
            type: string
        - in: query
          name: limit_runs
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 3
        - in: query
          name: limit_cards
          schema:
            type: integer
            minimum: 1
            maximum: 60
            default: 18
        - in: query
          name: limit_stories
          schema:
            type: integer
            minimum: 1
            maximum: 12
            default: 4
        - in: query
          name: limit_evidence_per_story
          schema:
            type: integer
            minimum: 1
            maximum: 8
            default: 3
      responses:
        '200':
          description: Watchlist trend with timeline and merged story groups
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistTrendResponse'
        '404':
          description: Not found
  /api/v1/watchlists/{watchlist_id}/briefing:
    get:
      operationId: getWatchlistBriefing
      parameters:
        - in: path
          name: watchlist_id
          required: true
          schema:
            type: string
        - in: query
          name: limit_runs
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 4
        - in: query
          name: limit_cards
          schema:
            type: integer
            minimum: 1
            maximum: 60
            default: 18
        - in: query
          name: limit_stories
          schema:
            type: integer
            minimum: 1
            maximum: 12
            default: 4
        - in: query
          name: limit_evidence_per_story
          schema:
            type: integer
            minimum: 1
            maximum: 8
            default: 3
      responses:
        '200':
          description: Unified watchlist briefing with summary, differences, and evidence drill-down
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WatchlistBriefingResponse'
        '404':
          description: Not found
...
  /api/v1/retrieval/search:
    post:
      operationId: retrievalSearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetrievalSearchRequest'
      responses:
        '200':
          description: Retrieval search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetrievalSearchResponse'
  /api/v1/retrieval/answer:
    post:
      operationId: retrievalAnswer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskAnswerRequest'
      responses:
        '200':
          description: Briefing-aware answer contract
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskAnswerResponse'
  /api/v1/retrieval/answer/page:
    post:
      operationId: retrievalAnswerPage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskPageRequest'
      responses:
        '200':
          description: Server-owned Ask page payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskPageResponse'
...
components:
  schemas:
...
    Platform:
      type: string
      description: Source platform key. `youtube` and `bilibili` are strong-supported video lanes; other string values are allowed for generic RSSHub/RSS substrate usage.
      examples: [youtube, bilibili, rsshub, generic]
...
    Watchlist:
      type: object
      required:
        [id, name, matcher_type, matcher_value, delivery_channel, enabled, created_at, updated_at]
      properties:
        id:
          type: string
        name:
          type: string
        matcher_type:
          type: string
          enum: [topic_key, claim_kind, platform, source_match]
        matcher_value:
          type: string
        delivery_channel:
          type: string
          enum: [dashboard, email]
        enabled:
          type: boolean
        created_at:
          type: string
        updated_at:
          type: string
    WatchlistUpsertRequest:
      type: object
      required: [name, matcher_type, matcher_value]
      properties:
        id:
          type: string
          nullable: true
        name:
          type: string
        matcher_type:
          type: string
          enum: [topic_key, claim_kind, platform, source_match]
        matcher_value:
          type: string
        delivery_channel:
          type: string
          enum: [dashboard, email]
          default: dashboard
        enabled:
          type: boolean
...
    WatchlistTrendRun:
      type: object
      required:
        [job_id, video_id, platform, title, created_at, matched_card_count, cards, topics, claim_kinds, added_topics, removed_topics, added_claim_kinds, removed_claim_kinds]
      properties:
        job_id:
          type: string
        video_id:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
        title:
          type: string
        source_url:
          type: string
          nullable: true
        created_at:
          type: string
        matched_card_count:
          type: integer
        cards:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistTrendCard'
        topics:
          type: array
          items:
            type: string
        claim_kinds:
          type: array
          items:
            type: string
        added_topics:
          type: array
          items:
            type: string
        removed_topics:
          type: array
          items:
            type: string
        added_claim_kinds:
          type: array
          items:
            type: string
        removed_claim_kinds:
          type: array
          items:
            type: string
    WatchlistMergedStory:
      type: object
      required:
        [id, story_key, headline, latest_created_at, matched_card_count, platforms, claim_kinds, source_urls, run_ids, cards]
      properties:
        id:
          type: string
        story_key:
          type: string
        headline:
          type: string
        topic_key:
          type: string
          nullable: true
        topic_label:
          type: string
          nullable: true
        latest_created_at:
          type: string
        matched_card_count:
          type: integer
        platforms:
          type: array
          items:
            type: string
        claim_kinds:
          type: array
          items:
            type: string
        source_urls:
          type: array
          items:
            type: string
        run_ids:
          type: array
          items:
            type: string
        cards:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistTrendCard'
    WatchlistTrendSummary:
      type: object
      required: [recent_runs, matched_cards, matcher_type, matcher_value]
      properties:
        recent_runs:
          type: integer
        matched_cards:
          type: integer
        matcher_type:
          type: string
        matcher_value:
          type: string
    WatchlistTrendResponse:
      type: object
      required: [watchlist, summary, timeline, merged_stories]
      properties:
        watchlist:
          $ref: '#/components/schemas/Watchlist'
        summary:
          $ref: '#/components/schemas/WatchlistTrendSummary'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistTrendRun'
        merged_stories:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistMergedStory'
...
    WatchlistBriefingSummary:
      type: object
      required: [overview, source_count, run_count, story_count, matched_cards, signals]
      properties:
        overview:
          type: string
        source_count:
          type: integer
        run_count:
          type: integer
        story_count:
          type: integer
        matched_cards:
          type: integer
        primary_story_headline:
          type: string
          nullable: true
        signals:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistBriefingSignal'
...
    WatchlistBriefingDifferences:
      type: object
      required:
        [added_topics, removed_topics, added_claim_kinds, removed_claim_kinds, new_story_keys, removed_story_keys]
      properties:
        latest_job_id:
          type: string
          nullable: true
        previous_job_id:
          type: string
          nullable: true
        added_topics:
          type: array
          items:
            type: string
        removed_topics:
          type: array
          items:
            type: string
        added_claim_kinds:
          type: array
          items:
            type: string
        removed_claim_kinds:
          type: array
          items:
            type: string
        new_story_keys:
          type: array
          items:
            type: string
        removed_story_keys:
          type: array
          items:
            type: string
        compare:
          allOf:
            - $ref: '#/components/schemas/WatchlistBriefingCompare'
          nullable: true
...
    WatchlistBriefingEvidence:
      type: object
      required: [stories, featured_runs]
      properties:
        suggested_story_id:
          type: string
          nullable: true
        stories:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistBriefingStoryEvidence'
        featured_runs:
          type: array
          items:
            $ref: '#/components/schemas/WatchlistBriefingRunEvidence'
    WatchlistBriefingResponse:
      type: object
      required: [watchlist, summary, differences, evidence]
      properties:
        watchlist:
          $ref: '#/components/schemas/Watchlist'
        summary:
          $ref: '#/components/schemas/WatchlistBriefingSummary'
        differences:
          $ref: '#/components/schemas/WatchlistBriefingDifferences'
        evidence:
          $ref: '#/components/schemas/WatchlistBriefingEvidence'
...
    RetrievalSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        top_k:
          type: integer
          minimum: 1
          maximum: 50
          default: 10
        mode:
          type: string
          enum: [keyword, semantic, hybrid]
          default: keyword
        filters:
          type: object
          default: {}
          additionalProperties: true
    RetrievalHit:
      type: object
      required: [job_id, video_id, platform, video_uid, source_url, kind, source, snippet, score]
      properties:
        job_id:
          type: string
          format: uuid
        video_id:
          type: string
          format: uuid
        platform:
          $ref: '#/components/schemas/Platform'
        video_uid:
          type: string
        source_url:
          type: string
        title:
          type: string
          nullable: true
        kind:
          type: string
        mode:
          type: string
          nullable: true
        source:
          type: string
          enum: [digest, transcript, outline, knowledge_cards, comments, meta]
        snippet:
          type: string
        score:
          type: number
    RetrievalSearchResponse:
      type: object
      required: [query, top_k, filters, items]
      properties:
        query:
          type: string
        top_k:
          type: integer
          minimum: 1
        filters:
          type: object
          additionalProperties: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/RetrievalHit'
    AskAnswerRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
        watchlist_id:
          type: string
          nullable: true
        story_id:
          type: string
          nullable: true
        top_k:
          type: integer
          minimum: 1
          maximum: 20
          default: 5
        mode:
          type: string
          enum: [keyword, semantic, hybrid]
          default: keyword
        filters:
          type: object
          default: {}
          additionalProperties: true
    AskAnswerContext:
      type: object
      required: [selection_basis, mode, filters, briefing_available]
      properties:
        watchlist_id:
          type: string
          nullable: true
        watchlist_name:
          type: string
          nullable: true
        story_id:
          type: string
          nullable: true
        selected_story_id:
          type: string
          nullable: true
        story_headline:
          type: string
          nullable: true
        topic_key:
          type: string
          nullable: true
        topic_label:
          type: string
          nullable: true
        selection_basis:
          type: string
          enum: [requested_story_id, query_match, suggested_story_id, first_story, none]
        mode:
          type: string
          enum: [keyword, semantic, hybrid]
        filters:
          type: object
          additionalProperties: true
        briefing_available:
          type: boolean
    AskAnswerBody:
      type: object
      required: [direct_answer, summary, confidence]
      properties:
        direct_answer:
          type: string
        summary:
          type: string
        reason:
          type: string
          nullable: true
        confidence:
          type: string
          enum: [grounded, limited]
...
    AskAnswerResponse:
      type: object
      required: [query, context, answer, changes, citations, evidence, fallback]
      properties:
        query:
          type: string
        context:
          $ref: '#/components/schemas/AskAnswerContext'
        selected_story:
          allOf:
            - $ref: '#/components/schemas/AskAnswerSelectedStory'
          nullable: true
        answer:
          $ref: '#/components/schemas/AskAnswerBody'
        changes:
          $ref: '#/components/schemas/AskAnswerChanges'
        citations:
          type: array
          items:
            $ref: '#/components/schemas/AskAnswerCitation'
        evidence:
          $ref: '#/components/schemas/AskAnswerEvidence'
        fallback:
          $ref: '#/components/schemas/AskAnswerFallback'
    AskPageRequest:
      type: object
      properties:
        query:
          type: string
          default: ''
        watchlist_id:
          type: string
          nullable: true
        story_id:
          type: string
          nullable: true
        topic_key:
          type: string
          nullable: true
        top_k:
          type: integer
          minimum: 1
          maximum: 20
          default: 6
        mode:
          type: string
          enum: [keyword, semantic, hybrid]
          default: keyword
        filters:
          type: object
          additionalProperties: true
          default: {}
    AskPageResponse:
      type: object
      required:
        [question, mode, top_k, context, answer_state, answer_confidence, citations, fallback_actions]
      properties:
        question:
          type: string
        mode:
          type: string
          enum: [keyword, semantic, hybrid]
        top_k:
          type: integer
        context:
          $ref: '#/components/schemas/AskAnswerContext'
        answer_state:
          type: string
          enum: [briefing_grounded, missing_context, briefing_unavailable, no_confident_answer]
        answer_headline:
          type: string
          nullable: true
        answer_summary:
          type: string
          nullable: true
        answer_reason:
          type: string
          nullable: true
        answer_confidence:
          type: string
          enum: [grounded, limited]
        story_change_summary:
          type: string
          nullable: true
        briefing:
          allOf:
            - $ref: '#/components/schemas/WatchlistBriefingResponse'
          nullable: true
        story_focus:
          allOf:
            - $ref: '#/components/schemas/AskAnswerSelectedStory'
          nullable: true
        selected_story:
          allOf:
            - $ref: '#/components/schemas/WatchlistBriefingStoryEvidence'
          nullable: true
        retrieval:
          allOf:
            - $ref: '#/components/schemas/RetrievalSearchResponse'
          nullable: true
        citations:
          type: array
          items:
            $ref: '#/components/schemas/AskAnswerCitation'
        fallback_reason:
          type: string
          nullable: true
        fallback_next_step:
          type: string
          nullable: true
        fallback_actions:
          type: array
          items:
            type: object
            required: [kind, label]
            properties:
              kind:
                type: string
                enum: [open_briefing, open_story, open_job, open_knowledge, open_search]
              label:
                type: string
              route:
                type: string
                nullable: true
...